Address warning about indexing with char type values based on ASCII characters (same problem as in r51103). Unlikely to be a problem in this case so just quell by casting to int. SCL git 512968e and 425a9c3.
svn:revision:51341
svn:branch:trunk
svn:account:n_reed

+++ src/other/step/include/express/object.h
@@ -88,10 +88,10 @@ extern SCL_EXPRESS_EXPORT struct Object * OBJ;
 /* macro function definitions */
 /******************************/
 
-#define OBJget_symbol(obj,type)     (*OBJ[type].get_symbol)(obj)
-#define OBJget_type(_type_)     (OBJ[_type_].type)
-#define OBJget_bits(_type_)     (OBJ[_type_].bits)
-#define OBJtype_is_oneof(_type_,class)  (OBJ[_type_].bits & (class))
+#define OBJget_symbol(obj,type)     (*OBJ[(int)type].get_symbol)(obj)
+#define OBJget_type(_type_)     (OBJ[(int)_type_].type)
+#define OBJget_bits(_type_)     (OBJ[(int)_type_].bits)
+#define OBJtype_is_oneof(_type_,class)  (OBJ[(int)_type_].bits & (class))
 #define OBJget_name(obj,type)       (OBJget_symbol(obj,type)->name)
 
 /***********************/
